remove donation policies and removed dead code
