+++ src/components/page/donor-dashboard/dashboard-view.tsx
@@ -3,8 +3,10 @@ import { DataTable } from "@/src/components/page/donor-dashboard/data-table";
 import ChartAreaInteractive from "@/src/components/page/donor-dashboard/chart-area-interactive";
 import { SectionCards } from "@/src/components/page/donor-dashboard/section-cards";
 import { SidebarInset, SidebarProvider } from "@/src/components/ui/sidebar";
-import { getDonationRowsForCurrentUser } from "@/src/lib/supabase/queries/get-donation-rows";
-import { getDonationSeriesForCurrentUser } from "@/src/lib/supabase/queries/get-donation-series";
+import {
+  getDonationRowsForCurrentUser,
+  getDonationSeriesForCurrentUser,
+} from "@/src/api/donor";
 
 export default async function DashboardView() {
   const rows = await getDonationRowsForCurrentUser();
